Move view to the right location
