+++ src/gui/components/record/view.tsx
